backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
redis==5.2.1
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.19
slowapi==0.1.9
